+++ charts/fastly-tls-operator/CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-n/a
+### Added
+
+* Address [issue](https://github.com/seatgeek/fastly-tls-operator/issues/1) where certificate private keys other than `algorithm: RSA` couldn't be decoded.
+* Add support for PKCS8 encoding.
 
 ## [0.1.0] - 2025-08-25